+++ packages/delegate/src/createRequest.ts
@@ -242,11 +242,7 @@ function projectArgumentValue(argValue: any, argType: GraphQLInputType): any {
       projectArgumentValue(item, argType.ofType),
     );
   }
-  if (
-    isInputObjectType(argType) &&
-    typeof argValue === 'object' &&
-    argValue !== null
-  ) {
+  if (isInputObjectType(argType) && typeof argValue === 'object') {
     const projectedValue: any = {};
     const fields = argType.getFields();
     for (const key in argValue) {
@@ -260,16 +256,14 @@ function projectArgumentValue(argValue: any, argType: GraphQLInputType): any {
     }
     return projectedValue;
   }
-  if (argValue != null) {
-    if (argType.name === 'Boolean') {
-      return Boolean(argValue);
-    }
-    if (argType.name === 'Int' || argType.name === 'Float') {
-      return Number(argValue);
-    }
-    if (argType.name === 'String') {
-      return String(argValue);
-    }
+  if (argType.name === 'Boolean') {
+    return Boolean(argValue);
+  }
+  if (argType.name === 'Int' || argType.name === 'Float') {
+    return Number(argValue);
+  }
+  if (argType.name === 'String') {
+    return String(argValue);
   }
   return argValue;
 }