+++ packages/logger/tests/legacyLogger.test.ts
@@ -9,12 +9,8 @@ export const _: MeshLogger = new LegacyLogger(null as any);
 
 function createTLogger(opts?: Partial<LoggerOptions>) {
   const writer = new MemoryLogWriter();
-  return [
-    LegacyLogger.from(
-      new Logger({ ...opts, writers: opts?.writers ? opts.writers : [writer] }),
-    ),
-    writer,
-  ] as const;
+  const writers = opts?.writers ?? [writer];
+  return [LegacyLogger.from(new Logger({ ...opts, writers })), writer] as const;
 }
 
 it('should correctly write legacy logger logs', () => {