+++ packages/fusion-runtime/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @graphql-mesh/fusion-runtime
 
+## 1.7.1
+### Patch Changes
+
+- Updated dependencies [[`dae3fe4`](https://github.com/graphql-hive/gateway/commit/dae3fe4fbb9c78598d1bbf91c2ed07ab1b90f05d)]:
+  - @graphql-tools/wrap@11.1.10
+  - @graphql-tools/delegate@12.0.10
+  - @graphql-tools/batch-execute@10.0.6
+  - @graphql-tools/federation@4.2.14
+  - @graphql-tools/stitch@10.1.14
+  - @graphql-tools/stitching-directives@4.0.16
+
+## 1.7.0
+### Minor Changes
+
+
+
+- [`5468882`](https://github.com/graphql-hive/gateway/commit/5468882253927113fa22030bceb265a31101792e) Thanks [@ardatan](https://github.com/ardatan)! - Expose `subschemas` in `handleFederationSupergraph` result
+
+## 1.6.11
+### Patch Changes
+
+- Updated dependencies [[`43aaa6a`](https://github.com/graphql-hive/gateway/commit/43aaa6ab8c1c64cae92e9cdff7016ebdf1afa756)]:
+  - @graphql-tools/delegate@12.0.9
+  - @graphql-tools/stitch@10.1.13
+  - @graphql-tools/federation@4.2.13
+  - @graphql-tools/stitching-directives@4.0.15
+  - @graphql-tools/wrap@11.1.9
+
+## 1.6.10
+### Patch Changes
+
+- Updated dependencies [[`c75fd0a`](https://github.com/graphql-hive/gateway/commit/c75fd0a64730ab07a3a7115748841025e312f335)]:
+  - @graphql-hive/logger@1.1.0
+  - @graphql-mesh/transport-common@1.0.15
+
 ## 1.6.9
 ### Patch Changes