+++ packages/stitching-directives/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphql-tools/stitching-directives
 
+## 4.0.16
+### Patch Changes
+
+- Updated dependencies [[`dae3fe4`](https://github.com/graphql-hive/gateway/commit/dae3fe4fbb9c78598d1bbf91c2ed07ab1b90f05d)]:
+  - @graphql-tools/delegate@12.0.10
+
+## 4.0.15
+### Patch Changes
+
+- Updated dependencies [[`43aaa6a`](https://github.com/graphql-hive/gateway/commit/43aaa6ab8c1c64cae92e9cdff7016ebdf1afa756)]:
+  - @graphql-tools/delegate@12.0.9
+
 ## 4.0.14
 ### Patch Changes